projects/rag-api-project/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0

# LLM & AI
langchain==0.1.6
langchain-openai==0.0.6
langchain-community==0.0.19
langchain-aws==0.1.0
openai==1.10.0
boto3==1.34.0

# Vector stores
faiss-cpu==1.7.4
pinecone-client==3.0.0

# Document processing
pypdf==4.0.0
python-docx==1.1.0
unstructured==0.12.0

# HTTP & async
httpx==0.26.0
aiofiles==23.2.1

# Monitoring & logging
python-json-logger==2.0.7
slowapi==0.1.9

# Testing
pytest==8.0.0
pytest-asyncio==0.23.0
pytest-cov==4.1.0
httpx==0.26.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
